requirements.txt
boto3>=1.40.0,<2.0.0
mcp[cli]>=1.18.0,<2.0.0
python-dotenv>=1.1.0,<2.0.0
starlette>=0.47.0,<1.0.0
uvicorn[standard]>=0.35.0,<1.0.0
